pyproject.toml
[project]
name = "wild-travelers"
version = "0.1.0"
description = "Répertoire interactif d'hébergements et restaurants en France"
authors = [
    {name = "Chinnawat Wisetwongsa",email = "binn1908@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.55.0,<2.0.0)",
    "pandas (>=2.2,<3.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "pydeck (>=0.9.1,<0.10.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "nltk (>=3.9.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
